+++ test/apply_private_method_test.go
@@ -18,7 +18,7 @@ func (*PrivateMethodStruct) doSomething() bool {
 }
 
 func TestApplyPrivate(t *testing.T) {
-	Convey("TestApplyPrivate", t, func() {
+	SkipConvey("TestApplyPrivate", t, func() {
         Convey("patch private method in the different package", func() {
             f := new(fake.PrivateMethodStruct)
             var s *fake.PrivateMethodStruct
